Split active and passive partition into separate files
As both actors grow, having a single file with both implementations
makes navigation unnecessarily hard.
